package/bcc: add optional luajit dependency
The luajit dependency was explicitely disabled when bcc has
been added to Buildroot.
Even if the python binding is probably more used, enable
luajit binding if luajit package is enabled in the defconfig.

Fix the luajit command line used to build bcc.o, otherwise
bcc.o is build for the host.

Signed-off-by: Romain Naour <romain.naour@gmail.com>
Cc: Qais Yousef <qais.yousef@arm.com>